techtest/regions/serializers.py
from .models import Region
from rest_framework import serializers


class RegionSerializer(serializers.ModelSerializer):
    class Meta:
        model = Region
        fields = ['pk', 'code', 'name']
